Add a minimal GlobalSettings accessor
Expose a `GlobalSettings` struct to access document-wide settings
information, such as scale factor or axis direction.

The current API is EXTREMELY BAREBONE, requiring the user to manually
specify the name of the properties they want to access.

This is in order to avoid blocking on any particular details. Favoring
further discussions about unit precisions or how to expose a proper
axis direction API for later.

See also:
- https://github.com/lo48576/fbxcel-dom/pull/7